fix: Removed resolved query param from swagger download script
When resolved=false is present it seems that some definitions in the swagger files are not downloaded. Instead, they're replaced with a link to the definition, however the link cannot be accessed.

By removing the parameter the complete swagger file will be downloaded with all definitions in place.

(cherry picked from commit 211fe88b1bdf06840b63e0ca99d80cce660f7269)